Add CI tests for Wasi on Android
Co-authored by: evacchi <evacchi@users.noreply.github.com>
