+++ .github/dependabot.yml
@@ -4,6 +4,10 @@ updates:
     directory: "/"
     schedule:
       interval: daily
+  - package-ecosystem: gradle
+    directory: "/android-tests"
+    schedule:
+      interval: weekly
   - package-ecosystem: "github-actions"
     directory: "/"
     schedule:
